requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
tqdm==4.66.1
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openpyxl==3.1.2
SQLAlchemy==2.0.27
urllib3==2.2.1
logging==0.4.9.6 